Renamed jsx files to js
