+++ client/app/components/admin.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import {hashHistory, Link} from 'react-router';
 import Alert from 'react-s-alert';
-import Loader from './loader.jsx';
-import LogoUpload from './logo_upload.jsx';
+import Loader from './loader';
+import LogoUpload from './logo_upload';
 
 class Admin extends React.Component {
