+++ client/app/components/article.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link, hashHistory} from 'react-router';
-import Loader from './loader.jsx';
+import Loader from './loader';
 import Alert from 'react-s-alert';
 
 class ViewArticle extends React.Component {
