+++ client/app/components/browse_archives.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Loader from './loader.jsx';
+import Loader from './loader';
 import {hashHistory} from 'react-router';
 import Alert from 'react-s-alert';
