+++ client/app/components/edit.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {hashHistory} from 'react-router';
 import Alert from 'react-s-alert';
-import Loader from './loader.jsx';
+import Loader from './loader';
 
 class EditArticle extends React.Component {
   constructor(props) {
