+++ client/app/components/loader.js
