+++ client/app/components/logo_upload.js
