+++ client/app/components/simple_article.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router';
-import Loader from './loader.jsx';
+import Loader from './loader';
 import Alert from 'react-s-alert';
 
 class SimpleArticle extends React.Component {
