+++ client/app/routes.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 
-import App from './components/app.jsx';
-import Home from './components/home.jsx';
-import Login from './components/login.jsx';
-import Article from './components/article.jsx';
-import NewArticle from './components/new.jsx';
-import EditArticle from './components/edit.jsx';
-import ArticleHistory from './components/history.jsx';
-import Search from './components/search.jsx';
-import Admin from './components/admin.jsx';
-import Setup from './components/setup.jsx';
-import EditTopic from './components/edit_topics.jsx';
-import EditUser from './components/edit_users.jsx';
+import App from './components/app';
+import Home from './components/home';
+import Login from './components/login';
+import Article from './components/article';
+import NewArticle from './components/new';
+import EditArticle from './components/edit';
+import ArticleHistory from './components/history';
+import Search from './components/search';
+import Admin from './components/admin';
+import Setup from './components/setup';
+import EditTopic from './components/edit_topics';
+import EditUser from './components/edit_users';
 
 
 export default function () {
